Initial commit
create a dotfiles repo for setting up new mac books easily. This can be
run on fresh installs or for updating my dotfiles and system settings.
However the "setup.sh" likely needs significant work to ensure clean
updates. I also need to add java versions to jenv in the setup but will
need to look more into scripting this before adding it to the setup.