Ensure the entire file is fetched before execution
Wrap the entire script in a block to ensure the full contents of script is fetched before ececution. We will no longer need the "verify_setup.sh" script as the script is from github so should be safe to execute and we know that it will have all the content loaded before execution.