Remove erroneous defaults
- some defaults cause issues when set on a fresh install. for instance
the "write com.apple.Passwords" "EnableMenuBarExtra" option only works
if the password app has been used before but this is likely not the case
on a fresh install so it's best to just have apple prompt the user to
add it to the menu bar.
- the "com.apple.AppleMultitouchTrackpad" "TrackpadThreeFingerDrag"
option seems to cause issues with the right click / two finger click
selection interaction. This setting is now a default for macOS anyways
so it is not needed.